pyproject.toml
[tool.poetry]
name = "treasury-orchestrator"
version = "0.1.0"
description = "Treasury Orchestrator for Bankers Bank"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "treasury_orchestrator"},
    {include = "bank_connector"},
    {include = "quantengine"},
    {include = "treasury_observability"},
]

[tool.poetry.dependencies]
requests = "^2.32"
httpx = "^0.27"
    python = ">=3.9,<4.0"
pydantic = ">=1.10,<2"
fastapi = "^0.110"
psycopg2-binary = "^2.9"
sqlmodel = "0.0.16"
# pydantic-xsd = "^0.3.2"  # Temporarily removed to resolve dependency issues
uvicorn = "^0.29"
redis = "^5"
kafka-python = "^2.0"
prometheus-client = "^0.20"
aiokafka = "^0.10"
testcontainers = {version = "^4.3", extras=["redpanda"]}
jsonschema = "^4.22"
python-dateutil = "^2.9.0.post0"
sqlalchemy = ">=2.0,<2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
requests-mock = "^1.11"
black = "^23.0"
isort = "^5.0"
mypy = "^1.0"
fakeredis = "^2.20"
pytest-asyncio = ">=0.23,<1.0"
pytest-env = "^1.1"
setuptools = "^80.9.0"

[tool.poetry.scripts]
treasury-cli = "treasury_orchestrator.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
